requirements.txt
httpx
python-dotenv
pyyaml